python-app/requirements.txt
psycopg2-binary
pgvector
flask
